split-files: reorganize util and keccak_table
