+++ hashes/zkevm-keccak/src/lib.rs
@@ -3,11 +3,11 @@
 
 use halo2_base::halo2_proofs;
 
+mod keccak_circuit;
+
+/// Keccak packed multi
+pub use keccak_circuit::keccak_packed_multi;
 /// Util
-pub mod util;
+pub use keccak_circuit::util;
 
 pub use keccak_circuit::KeccakCircuitConfig as KeccakConfig;
-
-// ---- refactored
-mod keccak_circuit;
-pub mod keccak_table;
